fix: restructure commands - move all the commands folders into commands folder.
