+++ commands/fs/add.js
@@ -1,7 +1,7 @@
 import { appendFile } from "node:fs/promises";
 import path from "node:path";
 
-import store from "../store.js";
+import store from "../../store.js";
 
 let currentDir = "";
