+++ commands/fs/rn.js
@@ -1,7 +1,7 @@
 import { rename } from "node:fs/promises";
 import path from "node:path";
 
-import store from "../store.js";
+import store from "../../store.js";
 
 let currentDir = "";
