+++ commands/nwd/cd.js
@@ -1,8 +1,8 @@
 import { lstat } from "node:fs/promises";
 import path from "node:path";
 
-import { currentDir } from "../cwd.js";
-import store from "../store.js";
+import { currentDir } from "../../cwd.js";
+import store from "../../store.js";
 
 export const cd = async (changeDirectoryTo) => {
   if (!changeDirectoryTo) {
