+++ commands/os/index.js
@@ -1,0 +1,6 @@
+import { commandsEmitter } from "../../emitter.js";
+import { os } from "./os.js";
+
+
+
+commandsEmitter.on("os", ([option]) => execute(() => os(option)));
