+++ emitter.js
@@ -14,8 +14,20 @@ export const EVENTS_LIST = [
   "hash",
   "compress",
   "decompress",
+  ".exit",
 ];
 
-export const commandsEmitter = new EventEmitter().on("error", (error) =>
-  console.log(error.message)
+export const execute = async (command) => {
+  try {
+    await command();
+  } catch (error) {
+    commandsEmitter.emit("error", error);
+  } finally {
+    commandsEmitter.emit("commandEnd");
+  }
+};
+
+export const commandsEmitter = new EventEmitter().on(
+  "error",
+  (error) => console.log(error.message)
 );
